updates for vtk multiaxis, offsets table, try/excpt clean/logging option

+++ src/qtpyvcp/widgets/display_widgets/vtk_backplot/base_backplot.py
@@ -52,7 +52,7 @@ class BaseBackPlot(object):
         if filename is None or not os.path.isfile(filename):
             self.canon = None
             self.notification.notification_dispatcher.setNotify("3D plot", "Can't load backplot, invalid file: {}".format(filename))
-            # raise ValueError("Can't load backplot, invalid file: {}".format(filename))
+            return
 
         self.last_filename = filename
 
@@ -76,18 +76,19 @@ class BaseBackPlot(object):
                 msg = gcode.strerror(result)
                 fname = os.path.basename(filename)
                 self.notification.notification_dispatcher.setNotify("3D plot", f"Error in {fname} line {seq - 1}\n{msg}")
-                # raise SyntaxError("Error in %s line %i: %s" % (fname, seq - 1, msg))
+                return
 
         except KeyboardInterrupt:
             # probably raised by an (AXIS, stop) comment in the G-code file
             # abort generating the backplot
-            pass
-        
-        except Exception as e:
-            LOG.warning(f"CANON ERROR {e}")
-        # clean up temp var file and the backup
-        os.unlink(self.temp_parameter_file)
-        os.unlink(self.temp_parameter_file + '.bak')
+            return
+        finally:
+            # Clean up temp var file and backup when present.
+            if os.path.isfile(self.temp_parameter_file):
+                os.unlink(self.temp_parameter_file)
+            bak_file = self.temp_parameter_file + '.bak'
+            if os.path.isfile(bak_file):
+                os.unlink(bak_file)
 
 
 if __name__ == "__main__":
